refactor: combine CruiseMenu
